requirements.txt
# Core dependencies
freqtrade>=2024.1
pandas>=2.0.0
pandas-ta-classic==0.3.78
ccxt>=4.0.0
python-telegram-bot>=20.0
Pillow>=10.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Development
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0
feedparser
